requirements.txt
gradio
langchain
langchain-community
chromadb
sentence-transformers
ollama
pydantic
python-dotenv
openai-whisper
gtts
pydub
numpy
pandas
langchain_openai
langchain_ollama
python-dotenv
langchain-huggingface
